multiple improvements and refactors regarding utility functions to read http responses data and create history items based on it

+++ cmd/passive.go
@@ -27,7 +27,7 @@ to quickly create a Cobra application.`,
 			fmt.Println("An ID needs to be provided")
 			os.Exit(0)
 		}
-		history, err := db.Connection.GetHistory(passiveHistoryID)
+		history, err := db.Connection.GetHistory(uint(passiveHistoryID))
 		if err != nil {
 			log.Panic().Err(err).Msg("Could not find a issue with the provided ID")
 			os.Exit(0)
